+++ updatable_ticker_library/lib/updatable_ticker.dart
@@ -1,1 +1,2 @@
 export 'src/updatable_ticker.dart';
+export 'src/led/updatable_led_ticker.dart';